software/MedianFunction/src/main/java/andrewmartin/lambda/Stats.java
package andrewmartin.lambda;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

public class Stats {
    private double median;

    public Stats(double median) {
        this.median = median;
    }

    public double getMedian() {
        return median;
    }

    public String toJson() {
        try {
            ObjectMapper objectMapper = new ObjectMapper();
            return objectMapper.writeValueAsString(this);
        } catch (JsonProcessingException e) {
            return "{}";
        }
    }
}
